Feature/enable memcached (#9)
* feat: Add Makefile with shortcuts for Docker tasks.

Add a Makefile with shortcuts for common Docker tasks.
- The Makefile searches for .env.dist files in the current directory and its subdirectories, and copies them to .env files.
- A shortcut to enable XDebug mode.
 - Update a specific client ID in the application's .env file.

* feat: Add memcached container to Docker Compose.

Added a new container, `app-memcached`, to the Docker Compose file to provide a Memcached caching service to system.
The container is configured with the memcached image, and the necessary environment variables are set via an environment file `memcached.env`.

* feat: Add caching exceptions.

Adding new exception classes, `CacheException` and `InvalidArgumentException`, to handle exceptions related to caching functionality.

* feat: Add `CacheAdapter` abstract class.

Adding `CacheAdapter` class implements the CacheInterface and provides common functionality for caching adapters.

* feat: Add `MemcachedAdapter` implementation.

Add `MemcachedAdapter` class to the project, which extends the abstract CacheAdapter class and implements the CacheInterface.

* feat: Add `RedisAdapter` class with prototype functions.

Add `RedisAdapter` class which extends CacheAdapter. It's not fully implemented yet, and currently has only prototype functions to be used for future work.

* feat: Updated `CacheFactory` to implement new caching adapters.

 Added functionality to the CacheFactory to support new caching adapters, based on the incoming adapter type.

* feat: Add Memcached environment configs.

Include Memcached env configs in .env.dist

* fix: Remove clientID data from Makefile.

The Makefile previously contained a function to set clientID per each composer up to make it easier in development stage.